+++ .gitattributes
@@ -1,2 +1,1 @@
 .git*		export-ignore
-.editorconfig	export-ignore
