+++ Dockerfile
@@ -4,6 +4,6 @@ COPY . /app/
 WORKDIR /app
 RUN make build
 RUN make test
-RUN php -S localhost:8888 server.php&sleep 2&&bats integration.bats
+RUN php -S localhost:8888 tests/server.php & sleep 2 && bats tests/integration.bats
 RUN make DESTDIR=/ install
 ENTRYPOINT ["/usr/bin/pkgstats"]
