+++ go.mod
@@ -1,0 +1,3 @@
+module pkgstats-cli
+
+go 1.15
