+++ internal/pacman/pacman.go
@@ -1,4 +1,4 @@
-package main
+package pacman
 
 import (
 	"context"
