+++ internal/pacman/pacman_test.go
@@ -1,4 +1,4 @@
-package main
+package pacman
 
 import (
 	"fmt"
@@ -19,6 +19,20 @@ func init() {
 		fmt.Println("https://mirror.pkgbuild.com/core/os/x86_64")
 		os.Exit(0)
 	}
+}
+
+var Mocks = make(map[string]func())
+
+func TestMain(m *testing.M) {
+	mockName := os.Getenv("TEST_MOCK")
+	if mockName != "" {
+		mock, ok := Mocks[mockName]
+		if ok {
+			mock()
+		}
+	}
+
+	os.Exit(m.Run())
 }
 
 func TestGetInstalledPackages(t *testing.T) {
