+++ internal/system/system.go
@@ -1,4 +1,4 @@
-package main
+package system
 
 import (
 	"io/ioutil"
