+++ internal/system/system_test.go
@@ -1,4 +1,4 @@
-package main
+package system
 
 import (
 	"fmt"
@@ -20,6 +20,19 @@ func init() {
 	}
 }
 
+var Mocks = make(map[string]func())
+
+func TestMain(m *testing.M) {
+	mockName := os.Getenv("TEST_MOCK")
+	if mockName != "" {
+		mock, ok := Mocks[mockName]
+		if ok {
+			mock()
+		}
+	}
+
+	os.Exit(m.Run())
+}
 func TestGetMachine(t *testing.T) {
 	system := System{}
 	system.uname = os.Args[0]
